separate chat density
