+++ packages/fluentui/react-northstar/src/components/Chat/chatDensity.ts
@@ -1,0 +1,2 @@
+export type ChatDensity = 'comfy' | 'compact';
+export const defaultChatDensity: ChatDensity = 'comfy';
