+++ packages/fluentui/react-northstar/src/themes/teams/components/Chat/chatMessageStyles.ts
@@ -1,7 +1,7 @@
 import { ComponentSlotStylesPrepared, ICSSInJSStyle } from '@fluentui/styles';
 import { isNil } from 'lodash';
 
-import { ChatDensity, defaultChatDensity } from '../../../../components/Chat/chatContext';
+import { ChatDensity, defaultChatDensity } from '../../../../components/Chat/chatDensity';
 import { chatMessageSlotClassNames, ChatMessageStylesProps } from '../../../../components/Chat/ChatMessage';
 import { pxToRem } from '../../../../utils';
 import { getBorderFocusStyles } from '../../getBorderFocusStyles';
